refactor(modals): split 1008-LOC modals.ts into directory
Split `src/ui/modals.ts` (1008 LOC, 7 sibling modal classes + 3 shared
type interfaces) into a 7-file directory `src/ui/modals/` following the
same pattern as PR #249 (history-modal) and PR #250 (query-engine):

- `index.ts` — re-export shim (17 LOC)
- `types.ts` — LintFixCallbacks, LintCounts, FixReportPhase (38 LOC)
- `suggest-modals.ts` — FileSuggestModal + FolderSuggestModal (68 LOC)
- `ConfirmModal-class.ts` (46 LOC)
- `LintReportModal-class.ts` (164 LOC)
- `IngestReportModal-class.ts` (154 LOC)
- `FixReportModal-class.ts` (54 LOC)
- `MultiFileSuggestModal-class.ts` (526 LOC)

Zero external caller changes — TypeScript directory import resolution
routes `./modals` → `./modals/index.ts` automatically. All 1616
existing tests pass unchanged.

Technical debt resolved:
- Inline `import(...)` type annotations → proper `import type`
- Stale comment in lint controller.ts updated
